// src/core.v2/domain/@shared/types/tenant-settings.type.ts
import { Currency } from 'features/settings/modules/accounting/modules/currencies/types/Currency';

export type TenantSettingsType = {
  id: number;
  externalId: string;
  companyName: string;
  address: string;
  countryCode: string;
  countryExternalCode: string;
  header: string;
  primaryPhoneNumber1: string;
  primaryPhoneNumber2: string;
  primaryEmail: string;
  localCurrencyId: string;
  localCurrencyExternalId: string;
  systemCurrencyId: string;
  systemCurrencyExternalId: string;
  openBalanceWithMinusSign: boolean;
  totalsAccuracyDigits: number;
  quantitiesAccuracyDigits: number;
  pricesAccuracyDigits: number;
  ratesAccuracyDigits: number;
  percentagesAccuracyDigits: number;
  measuringUnitsAccuracyDigits: number;
  queryAccuracyDigits: number;
  decimalSeparator: string;
  thousandsSeparator: string;
  displayCurrencyOnTheRight: boolean;
  roundingMethod: boolean;
  state: string;
  aliasName: string;
  addressType: string;
  streetNumber: string;
  dateOfIncorporation: string;
  globalLocationNumber: string;
  createdAt: string;
  updatedAt: string;
  createdBy: string;
  updatedBy: string;
  firstPullSyncAt: string;
  lastPullSyncAt: string;
  firstPushSyncAt: string;
  lastPushSyncAt: string;
  localCurrency: Currency;
  systemCurrency: Currency;
};